Track subnet refactor  (#2228)
* refactor subnet sync

* fix lint

* Remove redundancies (#2229)

* remove redundant code

* fix lint

* fix lint

* Convert Subnet in  1 command (#2230)

* one command deploy

* use default private key to deploy validator manager contract

* fix lint

* remove redundant code

* address comments

---------


---------
