+++ cmd/nodecmd/create_devnet.go
@@ -13,6 +13,8 @@ import (
 	"sync"
 	"time"
 
+	"github.com/ava-labs/avalanche-cli/pkg/node"
+
 	"golang.org/x/exp/slices"
 
 	"github.com/ava-labs/avalanche-cli/pkg/ansible"
@@ -155,7 +157,7 @@ func generateCustomGenesis(
 }
 
 func setupDevnet(clusterName string, hosts []*models.Host, apiNodeIPMap map[string]string) error {
-	if err := checkCluster(clusterName); err != nil {
+	if err := node.CheckCluster(app, clusterName); err != nil {
 		return err
 	}
 	inventoryPath := app.GetAnsibleInventoryDirPath(clusterName)
@@ -178,7 +180,7 @@ func setupDevnet(clusterName string, hosts []*models.Host, apiNodeIPMap map[stri
 	} else {
 		endpointIP = ansibleHosts[ansibleHostIDs[0]].IP
 	}
-	endpoint := getAvalancheGoEndpoint(endpointIP)
+	endpoint := node.GetAvalancheGoEndpoint(endpointIP)
 	network := models.NewDevnetNetwork(endpoint, 0)
 	network = models.NewNetworkFromCluster(network, clusterName)
