+++ cmd/nodecmd/ssh.go
@@ -10,6 +10,8 @@ import (
 	"strings"
 	"sync"
 
+	"github.com/ava-labs/avalanche-cli/pkg/node"
+
 	"github.com/ava-labs/avalanche-cli/pkg/ansible"
 	"github.com/ava-labs/avalanche-cli/pkg/cobrautils"
 	"github.com/ava-labs/avalanche-cli/pkg/constants"
@@ -72,7 +74,7 @@ func sshNode(_ *cobra.Command, args []string) error {
 	} else {
 		clusterNameOrNodeID := args[0]
 		cmd := strings.Join(args[1:], " ")
-		if err := checkCluster(clusterNameOrNodeID); err == nil {
+		if err := node.CheckCluster(app, clusterNameOrNodeID); err == nil {
 			// clusterName detected
 			if len(args[1:]) == 0 {
 				return printClusterConnectionString(clusterNameOrNodeID, clustersConfig.Clusters[clusterNameOrNodeID].Network.Kind.String())
@@ -219,7 +221,7 @@ func printClusterConnectionString(clusterName string, networkName string) error 
 
 // GetAllClusterHosts returns all hosts in a cluster including loadtest and monitoring hosts
 func GetAllClusterHosts(clusterName string) ([]*models.Host, error) {
-	if exists, err := checkClusterExists(clusterName); err != nil || !exists {
+	if exists, err := node.CheckClusterExists(app, clusterName); err != nil || !exists {
 		return nil, fmt.Errorf("cluster %s not found", clusterName)
 	}
 	clusterHosts, err := ansible.GetInventoryFromAnsibleInventoryFile(app.GetAnsibleInventoryDirPath(clusterName))
